refactor: split GitHubUtils into focused submodules
Move each API helper into its own file under libs/GitHubUtils/.
