+++ scripts/libs/GitHubUtils/index.ts
@@ -1,0 +1,11 @@
+import getEmployeeLogins from './getEmployeeLogins';
+import getLatestApprovers from './getLatestApprovers';
+import getRequiredApprovingReviewCount from './getRequiredApprovingReviewCount';
+import listPullRequestCommits from './listPullRequestCommits';
+
+export default {
+    getEmployeeLogins,
+    getLatestApprovers,
+    getRequiredApprovingReviewCount,
+    listPullRequestCommits,
+};
